feat: implement chip 8 instructions
* feat: implement SE, SNE, LD_I, JP, RND, LD_DT, LD_ST, ADD_I

* fix: add unit test for RND_Vx_byte

+++ arabica/cpu/cpu.hpp
@@ -12,7 +12,7 @@ class CPU : public noncopyable {
 public:
   constexpr static uint16_t REGISTER_COUNT  = 16;
   constexpr static uint16_t PC_START        = 0x0200;
-  constexpr static uint16_t DEFAULT_RATE_HZ = 60;
+  constexpr static uint8_t  DEFAULT_RATE_HZ = 60;
 
   CPU()  = default;
   ~CPU() = default;
